Relocate guarded route file
