+++ client/src/pages/CatalogPage/index.tsx
@@ -1,27 +1,33 @@
 import { createStyles, Grid, makeStyles, Theme } from '@material-ui/core';
 import React, { useEffect, useState } from 'react';
-import { withAuth } from '../../HOC/withAuth';
 import { SeriesTable } from '../../Tables/SeriesTable';
+import { withAuth } from '../../utils/withAuth';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
     pageGrid: {
-      height: "100%"
+      height: '100%',
     },
   })
 );
 
 const CatalogPage = () => {
   const classes = useStyles();
-  const [fullHeight, setFullHeight] = useState<number>(window.innerWidth > 600 ? window.innerHeight - 100 : window.innerHeight - 92)
+  const [fullHeight, setFullHeight] = useState<number>(
+    window.innerWidth > 600 ? window.innerHeight - 100 : window.innerHeight - 92
+  );
 
   useEffect(() => {
     const handleResize = () => {
-      setFullHeight(window.innerWidth > 600 ? window.innerHeight - 100 : window.innerHeight - 92);
+      setFullHeight(
+        window.innerWidth > 600
+          ? window.innerHeight - 100
+          : window.innerHeight - 92
+      );
     };
     window.addEventListener('resize', handleResize);
     return () => window.removeEventListener('resize', handleResize);
-  }, [])
+  }, []);
 
   return (
     <div style={{ height: fullHeight }}>
